backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==42.0.2
python-decouple==3.8
requests==2.31.0
